app/backend/requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
aiohttp==3.9.4; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
attrs==23.2.0; python_version >= '3.7'
azure-core==1.30.1; python_version >= '3.7'
azure-storage-blob==12.19.1; python_version >= '3.7'
boto3==1.18.44; python_version >= '3.8'
botocore==1.34.84; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==42.0.5; python_version >= '3.7'
dataclasses-json==0.6.4; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
einops==0.7.0; python_version >= '3.8'
environs==9.5.0; python_version >= '3.6'
fastapi==0.68.1; python_version >= '3.8'
filelock==3.13.4; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.3.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.60.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.19.0; python_version >= '3.8'
huggingface-hub==0.22.2; python_full_version >= '3.8.0'
idna==3.7; python_version >= '3.5'
isodate==0.6.1
jinja2==3.1.3; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.4.0; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
langchain==0.1.15; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.32; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.42; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.0.1; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.47; python_version < '4.0' and python_full_version >= '3.8.1'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.1; python_version >= '3.8'
minio==7.2.5
mpmath==1.3.0
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
networkx==3.3; python_version >= '3.10'
numpy==1.21.2; python_version >= '3.9'
openai==1.17.0; python_full_version >= '3.7.1'
orjson==3.10.0; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pandas==1.3.3; python_version >= '3.9'
pillow==10.3.0; python_version >= '3.8'
protobuf==5.26.1; python_version >= '3.8'
psycopg2-binary==2.9.1; python_version >= '3.7'
pyarrow==15.0.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pycryptodome==3.20.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.8.2; python_version >= '3.8'
pydantic-core==2.18.1; python_version >= '3.8'
pymilvus==2.4.0; python_version >= '3.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.0; python_version >= '3.8'
python-multipart==0.0.5; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
regex==2023.12.25; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
s3transfer==0.10.1; python_version >= '3.8'
safetensors==0.4.3; python_version >= '3.7'
scikit-learn==0.24.2; python_version >= '3.9'
scipy==1.13.0; python_version >= '3.9'
sentence-transformers==2.6.1; python_full_version >= '3.8.0'
setuptools==69.5.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==1.4.23; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
sympy==1.12; python_version >= '3.8'
tenacity==8.2.3; python_version >= '3.7'
threadpoolctl==3.4.0; python_version >= '3.8'
tokenizers==0.15.2; python_version >= '3.7'
torch==1.9.0
tqdm==4.66.2; python_version >= '3.7'
transformers==4.9.2; python_full_version >= '3.8.0'
typing-extensions==4.11.0; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
ujson==5.9.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.10'
uvicorn==0.15.0; python_version >= '3.8'
websockets==12.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
spacy==3.1.3
